update billinginfo for nested addresses
