+++ components/donate/Billing.tsx
@@ -11,6 +11,7 @@ import {
 import { BillingInfo } from "../../src/types/BillingInfo";
 import styles from "./AddressForm.module.scss";
 import StateAutocomplete from "./StateAutocomplete";
+import AddressForm from "./AddressForm";
 
 type BillingProps = {
 	formData: BillingInfo;
@@ -28,64 +29,12 @@ const Billing = ({ formData, setFormData }: BillingProps) => {
 		<div>
 			<h2>Billing Information</h2>
 			<div className={styles.address}>
-				<TextField
-					id="first-name"
-					label="First Name"
+				<AddressForm
+					value={formData.billingAddress}
+					onChange={(address) => {
+						setFormData({ ...formData, billingAddress: address });
+					}}
 					required
-					sx={{ gridColumn: "span 2" }}
-					variant="outlined"
-					value={formData.firstName}
-					onChange={handleChange("firstName")}
-				/>
-				<TextField
-					id="last-name"
-					label="Last Name"
-					required
-					sx={{ gridColumn: "span 2" }}
-					variant="outlined"
-					value={formData.lastName}
-					onChange={handleChange("lastName")}
-				/>
-
-				<TextField
-					id="street-address"
-					label="Address"
-					required
-					sx={{ gridColumn: "span 4", marginTop: "1em" }}
-					variant="outlined"
-					value={formData.streetAddress}
-					onChange={handleChange("streetAddress")}
-				/>
-				<TextField
-					id="apt-suite"
-					label="Apt/Suite/Bldg"
-					sx={{ gridColumn: "span 2" }}
-					variant="outlined"
-					value={formData.aptOrSuite}
-					onChange={handleChange("aptOrSuite")}
-				/>
-				<TextField
-					id="city"
-					label="City"
-					required
-					sx={{ gridColumn: "span 2" }}
-					variant="outlined"
-					value={formData.city}
-					onChange={handleChange("city")}
-				/>
-				<StateAutocomplete
-					value={formData.state}
-					onChange={(newValue) => setFormData({ ...formData, state: newValue })}
-					required
-				/>
-				<TextField
-					id="zip-code"
-					label="Zip Code"
-					required
-					sx={{ gridColumn: "span 2" }}
-					variant="outlined"
-					value={formData.zip}
-					onChange={handleChange("zip")}
 				/>
 				<TextField
 					id="email"
