+++ src/types/Address.ts
@@ -1,0 +1,9 @@
+export interface Address {
+	firstName: string;
+	lastName: string;
+	streetAddress: string;
+	aptOrSuite: string;
+	city: string;
+	state: string;
+	zip: string;
+}
